Replace custom Webflow client with webflow-rb gem integration
